Refactor  cookingSchedule by seperating date picker and Date range picker
